% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI.R
\name{calculateGUI}
\alias{calculateGUI}
\alias{GUI}
\title{A graphical user interface for the package mdir.logrank}
\usage{
calculateGUI()
}
\description{
This function provides a graphical user interface for calculating
multiple-direction logrank test and the corresponding p-values
(based on \eqn{\chi^2}-approximation as well as on permutation).
}
