% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_mdirlogrank.R
\name{mdir.logrank.shiny}
\alias{mdir.logrank.shiny}
\title{A shiny app for the package mdir.logrank}
\usage{
mdir.logrank.shiny()
}
\description{
This function provides a shiny app for calculating the
multiple-direction logrank test for the two-sided and the one-sided
testing problem.
}
