% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdir.onesided.R
\name{mdir.onesided}
\alias{mdir.onesided}
\title{Two-sample multiple-direction log rank test for stochastic ordered alternatives}
\usage{
mdir.onesided(data, group1, rg = list(c(0, 0), c(0, 4), c(4, 0)),
  wild = "rade", iter = 10000, dig_p = 3, dig_stat = 3)
}
\arguments{
\item{data}{A data.frame, list or environment containing the variables \code{time},
\code{event} (with values 0 for censored and 1 for uncensored) and \code{group}.}

\item{group1}{The name or the coding for the first group in the data set}

\item{rg}{A list containing the exponents \code{c(r, g)} of the directions
 \eqn{w(x) = x^r (1-x)^g}. Both exponents r,g need to be natural numbers including 0.
Default is \code{list(c(0, 0), c(0, 4), c(4, 0))} corresponding to the choice of the
proportional, early and late direction/weight.}

\item{wild}{The wild bootstrap approach used for estimating the p-value. The Rademacher
(\code{rade}, default), the normal distribution (\code{norm}) or the centred
Poisson distribution (\code{pois}) approach can be selected.}

\item{iter}{The number of iteration used for calculating the wild bootstrap p-value.
The default option is 10000.}

\item{dig_p}{The p-values are rounded to \code{dig_p} digits, the default is 3.}

\item{dig_stat}{The test statistic is rounded to \code{dig_stat} digits, the default is 3.}
}
\value{
An  \code{mdirone} object containing the following components:
\item{Descriptive}{The directions used and whether the directions specified by the user were
   linearly independent.}
 \item{p.value}{The p-value of the one-sided multiple-direction logrank test using the
   the using the permutation approach (Perm.).}
 \item{wild}{The wild bootstrap approach which was used}
 \item{stat}{Value of the one-sided multiple-direction logrank statistic.}
 \item{rg}{A list containing the exponents of the direction considered in the statistical analysis.}
 \item{group1}{The name of the first group.}
 \item{indep}{logical. Were the directions specified by the user linearly independent?}
 \item{iter}{The number of iterations used for calculating the wild bootstrap p-value.}
}
\description{
The mdir.onesided function calculates the multiple-direction logrank
statistic for (one-sided) stochastic ordered alternatives and
its p-value  based on a wild bootstrap approach
}
\details{
The function provides the multiple-direction logrank statistic for
  the two sample one-sided testing problem of stochastic ordering within right-censored survival data.
  The null hypothesis \eqn{H:F1=F2} is tested against the one-sided alternative \eqn{K:F1 \ge F2,
  F1 \neq F2}. The first group corresponding to \eqn{F1} can be specified
  by the argument \code{group1}. An arbitrary amount of directions/weights of the form
  \eqn{w(x) = x^r (1-x)^g} for natural numbers r,g (including 0) can be chosen in the list
  \code{rg}. The multiple-direction onsided logrank test needs linearly independent directions.
  A check for this is implemented. If the directions chosen by the user are
  linearly dependent then a subset consisting of linearly independent directions
  is selected automatically.

  The \code{mdir.logrank} function returns the test statistic and the p-value
  based on a wild bootstrap procedure \code{wild}.
}
\examples{
library(coin)
data(GTSG)
out <- mdir.logrank(data = GTSG, group1 = "Chemotherapy+Radiation")

## Detailed information:
summary(out)

}
\references{
Ditzhaus, M., Pauly, M. (2018). ????
}
\seealso{
\code{\link[mdir.logrank]{mdir.onesided}}
}
